Remove Zeek testcase from Zmq message roundtrips
